+++ src/com/lcpan/dao/SalesRecordDAO.java
@@ -14,7 +14,7 @@ public interface SalesRecordDAO {
 	
 	public void deleteSalesRecord(String orderNumber);
 	
-	public int insertGetSalesRecord();
+	public long insertGetSalesRecord();
 	public void insertSalesRecord(String date, String orderNumber, String productNo, String amount, String price,
 			String discount, String totalPrice, String gender, String number);
 	public boolean delPay(String productNo);
